fixing test and spliting Struct
